requirements.txt
PyMuPDF==1.23.8
Pillow==10.1.0
boto3==1.34.20
psycopg2-binary==2.9.9
python-dotenv==1.0.0
